rewrote rescuepoint into alars_recover
