+++ behaviours/alars/setup.py
@@ -23,7 +23,7 @@ setup(
     tests_require=['pytest'],
     entry_points={
         'console_scripts': [
-            "rescuepoint_server= alars.rescuepoint_server:main",
+            "alars_recover_action_server = alars.alars_recover_action_server:main",
             "alars_search_action_server = alars.alars_search_action_server:main",
             "alars_localize_action_server = alars.alars_localize_action_server:main",
         ],
