+++ scripts/smarc_bringups/scripts/dji_bringup.sh
@@ -115,7 +115,7 @@ tmux send-keys "ros2 run alars alars_localize_action_server --ros-args -r __ns:=
 -p tracking_tolerance:=0.1 -p tracking_aggressiveness:=3.0" C-m
 
 tmux select-pane -t $SESSION:1.2
-tmux send-keys "echo 'This will be alars-recover'" C-m
+tmux send-keys "ros2 run alars alars_recover_action_server --ros-args -r __ns:=/$ROBOT_NAME -p use_sim_time:=$USE_SIM_TIME" C-m
 
 tmux select-pane -t $SESSION:1.3
 tmux send-keys "echo 'This will be alars-checkload'" C-m
